GH-40 Move the tree component to scenes
